move source files in module directory
